extract logger
