+++ exo-go/src/application/initializer_test.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/Originate/exosphere/exo-go/src/application"
 	"github.com/Originate/exosphere/exo-go/src/dockercompose"
+	"github.com/Originate/exosphere/exo-go/src/logger"
 	"github.com/Originate/exosphere/exo-go/src/types"
 	"github.com/Originate/exosphere/exo-go/src/util"
 	"github.com/Originate/exosphere/exo-go/test_helpers"
@@ -39,7 +40,7 @@ var _ = Describe("Initializer", func() {
 		appConfig, err := types.NewAppConfig(appDir)
 		Expect(err).NotTo(HaveOccurred())
 		_, pipeWriter := io.Pipe()
-		mockLogger := application.NewLogger([]string{}, []string{}, pipeWriter)
+		mockLogger := logger.NewLogger([]string{}, []string{}, pipeWriter)
 		initializer, err := application.NewInitializer(appConfig, mockLogger, appDir, homeDir)
 		Expect(err).NotTo(HaveOccurred())
 		err = initializer.Initialize()
