+++ exo-go/src/logger/logger.go
@@ -1,4 +1,4 @@
-package application
+package logger
 
 import (
 	"fmt"
